🚚 Move lxml-based `MigXmlReader` to a separate module
